Refactored sorting to be more streamlined
It's @Dash from Stackoverflow. I've been thinking about the problem for a while and came up with a solution for the x-axis not lining up. Instead of sorting the datapoints and plotting that, it collects all the values for each node in each run, and then plots those, making sure that each node has one x value.